Fixed many more compile warnings. With a few locally suppressed exceptions, we are now down only to warnings from external libraries. (mostly cpp_redis).

+++ Plugins/JVM/Util.h
@@ -25,5 +25,5 @@ inline uint64_t Measure(Lambda code)
 
     // assert(diff >= 0);
 
-    return diff;
+    return static_cast<uint64_t>(diff);
 }
